Refactor the original AutoSite Legacy code a bit
In order of when I bothered to add them to the list:

 - Default template now exists as a constant with an uppercase SNAKE_CASE name
 - `dirname` function, which was kind of awkwardly placed, is now outside of the `main` function blob
 - Remove `auto` option, since it didn't actually do anything
 - Remove `prettify` option, since it can be done as a plugin if you really need it (and I would have had to pass it to a refactored file handler)
 - Replace `bcolors` (which is GPL like the rest of blender) with a new ANSI color library/class thing I wrote from scratch
    - While changing over to new ANSI color library/class thing, standardize on green as "success" and yellow as "warning" (as opposed to the weird thing the original had where success was blue except for the two places it was green)
    - Also, add "boring" option to turn off colors.
 - Split the pre-processing checks for directory structure and template existence into their own function
 - Split the `in/` directory walker into its own function
    - Refactor the check that the file being processed in the `in/` dir isn't Thumbs.db or .DS_Store to handle easy extension (in case any new meta files we don't want processed become a thing).
 - Split the file processing into its own function

Also probably some other stuff I haven't bothered to add to this list. Basically, I took the stance of "if I were writing this, would I have split this into its own function" and erred on the side of doing so (I probably would have kept the in directory walker in main, but I figured it could possibly be easier to understand split into its own function).

I also didn't bother implementing any of the other intended changes between hamilton and AutoSite Legacy (save for the new shiny coat of paint that is the hamilton branding). Do one thing and do it well, as they say.